LayoutTests/imported/w3c/web-platform-tests/content-security-policy/generic/generic-0_10_1.sub-expected.txt

CONSOLE MESSAGE: Refused to load http://www.localhost:8801/content-security-policy/generic/unreached.js because it does not appear in the script-src directive of the Content Security Policy.
implicit port number matching fails with a different port


Harness Error (TIMEOUT), message = null

PASS Prevents access to external scripts.
NOTRUN Should fire violation events for every failed violation
